Cleanup of obsolete code
