--- src/main/java/com/phasmidsoftware/visitor/JMessage.java
@@ -1,7 +1,0 @@
-package com.phasmidsoftware.visitor;
-
-public enum JMessage {
-    PRE,
-    POST,
-    IN
-}
