--- src/main/scala/com/phasmidsoftware/visitor/DfsVisitor.scala
@@ -1,482 +1,0 @@
-package com.phasmidsoftware.visitor
-
-import com.phasmidsoftware
-import com.phasmidsoftware.visitor.DfsVisitor.recurseWithInVisit
-
-/**
- * A specialized visitor implementing recursive traversal and processing logic.
- * Essentially, this method performs depth-first-search on a tree structure,
- * where a tree can, of course, be a subgraph of a graph.
- *
- * This case class extends `AbstractMultiVisitor` and is designed to handle recursive
- * visits of elements of type `X`, using the provided function `children` to generate sub-elements
- * for recursion. It is initialized with a map linking specific `Message` instances
- * to corresponding `Appendable[X]` objects.
- * The `Message` types supported by this method are: `Pre`, `In`, and `Post`.
- *
- * @param map a map associating `Message` types with `Appendable[X]` instances
- *            to manage state during the recursion process.
- * @param children a function of type `X => Seq[X]` that defines how to retrieve
- *            sub-elements (children) for recursive processing from a given instance of `X`.
- * @tparam X the type of elements being visited and processed recursively.
- */
-case class DfsVisitor[X](map: Map[Message, Appendable[X]], children: X => Seq[X]) extends AbstractMultiVisitor[X](map) with Dfs[X, DfsVisitor[X]] {
-  /**
-   * Recursively processes an element of type `X` using a depth-first traversal strategy.
-   * This method updates the internal state of the `DfsVisitor` at each step of the recursion,
-   * applying pre-, in-, and post-visit modifications as determined by the `Pre`-, `In`- and `Post`-messages.
-   *
-   * @param x the element of type `X` to be processed and traversed recursively.
-   * @return a new instance of `DfsVisitor` with the updated state after processing
-   *         and visiting the provided element and its sub-elements.
-   */
-  def dfs(x: X): DfsVisitor[X] = {
-    def performRecursion(xs: Seq[X], visitor: DfsVisitor[X]) =
-      if (xs.length <= 2 && map.contains(In))
-        DfsVisitor.recurseWithInVisit(visitor, x, xs)
-      else
-        xs.foldLeft(visitor)(_ dfs _)
-
-    // First do a pre-visit, then a "SelfVisit", next perform the recursion, finally do a post-visit
-    performRecursion(children(x), visit(Pre)(x).visit(SelfVisit)(x)).visit(Post)(x)
-  }
-
-  /**
-   * Make a visit, with the given message and `X` value, on this `Visitor` and return a new `Visitor`.
-   *
-   * This method defines the behavior for handling a `Message` in the context
-   * of the Visitor pattern. The implementation of this method should use the provided
-   * message and state to determine the next state and return the appropriate `Visitor`.
-   *
-   * @param msg the message to be processed by the visitor
-   * @param x   the current state or context associated with the visitor
-   * @return a new `Visitor[X]` instance that represents the updated state after processing the message
-   */
-  override def visit(msg: Message)(x: X): DfsVisitor[X] =
-    super.visit(msg)(x).asInstanceOf[DfsVisitor[X]]
-
-  /**
-   * Creates a new `Visitor` instance with the provided updated mapAppendables.
-   *
-   * This method is used to update the internal state of the Visitor by creating
-   * a new instance with the modified mappings from `Message` to `Appendable`.
-   *
-   * @param map a map containing updated associations of `Message` to `Appendable[X]`
-   * @return a new `Visitor[X]` instance that reflects the updated mapAppendables
-   */
-  def unit(map: Map[Message, Appendable[X]]): DfsVisitor[X] =
-    copy(map = map)
-}
-
-/**
- * Contains utility methods for processing recursive logic within the `DfsVisitor` class.
- *
- * This object provides an internal helper function used to facilitate specific
- * recursive operations with handling of `In` messages during traversal.
- */
-object DfsVisitor {
-
-  /**
-   * Creates a new instance of `DfsVisitor` to perform a depth-first traversal of elements of type `X`.
-   * The traversal behavior is determined by the provided message, journal, and recursive function.
-   *
-   * @param message the `Message` type determining the current phase of the visit (e.g., Pre, In, Post).
-   * @param journal an `Appendable[X]` instance to manage and store the visited elements during traversal.
-   * @param children a function of type `X => Seq[X]` that defines how to obtain the children of a given element.
-   * @return a new instance of `DfsVisitor[X]` initialized with the provided parameters.
-   */
-  def create[X](message: Message, journal: Appendable[X], children: X => Seq[X]): DfsVisitor[X] =
-    new DfsVisitor(Map(message -> journal), children)
-
-  /**
-   * Creates a `DfsVisitor` instance configured for depth-first traversal with a pre-visit strategy
-   * and an empty `QueueJournal` for maintaining state.
-   *
-   * This method initializes the visitor to process elements of type `X` using the provided function `children`
-   * to generate child elements for recursion. During traversal, elements are pre-visited before their children.
-   *
-   * @param children a function of type `X => Seq[X]` that defines how to retrieve
-   *          sub-elements (children) for recursive processing from a given instance of `X`
-   * @tparam X the type of the elements to be visited and processed recursively
-   * @return a `DfsVisitor[X]` instance configured with the pre-visit strategy
-   *         and an empty `QueueJournal` for managing traversal state
-   */
-  def createPreQueue[X](children: X => Seq[X]): DfsVisitor[X] =
-    create(Pre, QueueJournal.empty[X], children)
-
-  /**
-   * Creates a `DfsVisitor` instance configured to use a post-order traversal
-   * with an empty `QueueJournal` for maintaining the traversal state.
-   *
-   * The returned visitor employs the provided function `children` to determine
-   * the sub-elements (children) of each element during the traversal process.
-   *
-   * @param children a function of type `X => Seq[X]` that specifies how to retrieve
-   *          the children of a given element of type `X` for the post-order traversal.
-   * @tparam X the type of elements to be processed during the depth-first traversal.
-   * @return a new instance of `DfsVisitor[X]` configured for post-order traversal
-   *         with a `QueueJournal` to handle traversal state.
-   */
-  def createPostQueue[X](children: X => Seq[X]): DfsVisitor[X] =
-    create(Post, QueueJournal.empty[X], children)
-
-  /**
-   * Creates a `DfsVisitor` instance that uses `Pre`-ordering and a `ListJournal` for traversal state.
-   *
-   * This method initializes a depth-first search visitor configured to perform
-   * a pre-order recursive traversal of a structure. The visitor uses the provided
-   * function to generate the children (sub-elements) for each visited element
-   * and utilizes a `ListJournal` to manage the traversal state.
-   *
-   * @param children a function of type `X => Seq[X]` that takes an element of type `X` and returns
-   *          a sequence of child elements. This function is used to derive the recursive
-   *          structure that the visitor will traverse.
-   * @return a `DfsVisitor` instance configured for pre-order traversal with a `ListJournal`.
-   * @tparam X the type of the elements to be processed during the recursive traversal.
-   */
-  def createPreStack[X](children: X => Seq[X]): DfsVisitor[X] =
-    create(Pre, ListJournal.empty[X], children)
-
-  /**
-   * Creates a depth-first traversal visitor that uses a post-order traversal strategy
-   * and employs a stack structure (`ListJournal`) for maintaining the state during recursion.
-   *
-   * @param children a function of type `X => Seq[X]` which, given an element of type `X`,
-   *          produces a sequence of its child elements for recursive traversal.
-   * @tparam X the type of elements being visited and processed recursively.
-   * @return an instance of `DfsVisitor[X]` configured for post-order traversal
-   *         using a stack-based journal (`ListJournal`).
-   */
-  def createPostStack[X](children: X => Seq[X]): DfsVisitor[X] =
-    create(Post, ListJournal.empty[X], children)
-
-  /**
-   * Performs a recursive traversal with specific handling for "In" messages.
-   * This method processes the provided element `x` within a recursive structure
-   * and updates the visitor state accordingly, ensuring it supports the case
-   * of sequences with at most two elements.
-   *
-   * CONSIDER should we pass in the dfs function as a parameter?
-   *
-   * @param visitor the current `DfsVisitor` instance that maintains
-   *                the state of the recursive traversal.
-   * @param x       the element of type `X` that is being processed and traversed.
-   * @param xs      a sequence of type `X` representing the sub-elements related to `x`.
-   *                This sequence must contain at most two elements.
-   */
-  private def recurseWithInVisit[X](visitor: DfsVisitor[X], x: X, xs: Seq[X]) = {
-    require(xs.length <= 2, "xs must contain at most two elements")
-    val visitor1 = xs.headOption.fold(visitor)(visitor.dfs)
-    val visitor2 = visitor1.visit(In)(x)
-    xs.lastOption.fold(visitor2)(visitor2.dfs)
-  }
-}
-
-/**
- * A case class for managing depth-first traversal using the Visitor design pattern based on a key-value pair.
- *
- * `DfsVisitorMapped` combines a custom traversal strategy with visitor behavior,
- * allowing for recursive traversal of elements (`K`) while applying transformations (`fulfill`)
- * to produce associated values (`Option[V]`). The traversal is guided by the map of `Message` to `Appendable`
- * and the `children` function that specifies the structure to navigate.
- *
- * @param map      a mapping from `Message` instances to `Appendable[(K, V)]`, defining how messages are handled.
- * @param fulfill  a function that takes an element of type `K` and returns an appropriate value of type `Option[V]`
- * @param children a function that takes an element of type `K` and returns a sequence of its child elements to traverse.
- * @tparam K the type of the keys or elements being traversed.
- * @tparam V the type of the associated values produced during the traversal process.
- */
-case class DfsVisitorMapped[K, V]
-(map: Map[Message, Appendable[(K, Option[V])]], fulfill: K => Option[V], children: K => Seq[K]) extends
-  AbstractVisitorMappedWithChildren[K, K, V](map, _ => fulfill, children)
-  with Dfs[K, DfsVisitorMapped[K, V]] {
-  /**
-   * Performs a depth-first traversal starting from the provided key `h`.
-   * This method applies visitor operations in a specific sequence: pre-visit, self-visit, recursive traversal, and post-visit.
-   *
-   * Depending on the conditions of the children of the key, the traversal may apply special in-visit logic if there are exactly two children.
-   * Returns an updated `DfsVisitorMapped` instance reflecting the changes made during the traversal.
-   *
-   * @param k the starting key for the depth-first search operation
-   * @return a new `DfsVisitorMapped[K, V]` instance after performing the DFS traversal from the given key
-   */
-  def dfs(k: K): DfsVisitorMapped[K, V] = {
-    val kv: (K, Option[V]) = keyValuePair(k)
-
-    def performRecursion(xs: Seq[K], visitor: DfsVisitorMapped[K, V]) =
-      if (xs.length <= 2 && map.contains(In))
-        DfsVisitorMapped.recurseWithInVisit(visitor, k, kv._2, xs)
-      else
-        xs.foldLeft(visitor)(_ dfs _)
-
-    // First do a pre-visit, then a "SelfVisit", next perform the recursion, finally do a post-visit
-    performRecursion(children(k), visit(Pre)(kv).visit(SelfVisit)(kv)).
-      visit(Post)(kv)
-  }
-
-  /**
-   * Make a visit, with the given message and `(K, V)` value, on this `Visitor` and return a new `Visitor`.
-   *
-   * This method defines the behavior for handling a `Message` in the context
-   * of the Visitor pattern. The implementation of this method should use the provided
-   * message and state to determine the next state and return the appropriate `Visitor`.
-   *
-   * @param msg the message to be processed by the visitor
-   * @param kv  the current state or context associated with the visitor
-   * @return a new `DfsVisitorMapped[K, V]` instance that represents the updated state after processing the message
-   */
-  override def visit(msg: Message)(kv: (K, Option[V])): DfsVisitorMapped[K, V] =
-    super.visit(msg)(kv).asInstanceOf[DfsVisitorMapped[K, V]]
-
-  /**
-   * Creates a new `Visitor` instance with the provided updated mapAppendables.
-   *
-   * This method is used to update the internal state of the Visitor by creating
-   * a new instance with the modified mappings from `Message` to `Appendable`.
-   *
-   * @param map a map containing updated associations of `Message` to `Appendable[(K, V)]`
-   * @return a new `DfsVisitorMapped[K, V]` instance that reflects the updated mapAppendables
-   */
-  def unit(map: Map[Message, Appendable[(K, Option[V])]]): DfsVisitorMapped[K, V] =
-    copy(map = map)
-}
-
-/**
- * Companion object for the `DfsVisitorMapped` class, incorporating utility methods to manage
- * depth-first traversal operations with a visitor-centric approach.
- */
-object DfsVisitorMapped {
-
-  /**
-   * Creates a new instance of `DfsVisitorMapped` to manage depth-first traversal using a visitor pattern.
-   *
-   * @param message  the `Message` instance representing the context or phase of the traversal
-   * @param journal  the `Appendable` structure for recording visited elements as pairs, each having type `(K, Option[V])`
-   * @param fulfill  a function that takes an element of type `K` and returns an appropriate value of type `Option[V]`
-   * @param children a function that takes an element of type `K` and returns its sequence of child elements to traverse
-   * @return a new instance of `DfsVisitorMapped[K, V]` configured with the specified message, appendable structure, and functions
-   */
-  def create[K, V](message: Message, journal: AbstractMapJournal[K, Option[V]], fulfill: K => Option[V], children: K => Seq[K]): DfsVisitorMapped[K, V] =
-    new DfsVisitorMapped(Map(message -> journal), fulfill, children)
-
-  /**
-   * Creates a `DfsVisitorMapped` instance configured for depth-first traversal using a pre-visit phase
-   * and a queue-based mechanism for recording visited elements.
-   *
-   * This method sets up the traversal with a pre-order strategy, utilizing a queue as the journal
-   * to manage visited elements during the traversal process. The functions provided define the mapping
-   * of an element to a value and how the children of each element are determined for recursive traversal.
-   *
-   * @param fulfill a function that takes an element of type `K` and returns an appropriate value of type `Option[V]`
-   * @param children a function that takes an element of type `K` and returns its sequence of child elements to traverse
-   * @return a new `DfsVisitorMapped[K, V]` instance configured for pre-order traversal using a queue
-   */
-  def createPreMapJournal[K, V](fulfill: K => Option[V], children: K => Seq[K]): DfsVisitorMapped[K, V] =
-    create(Pre, MapJournal.empty, fulfill, children)
-
-  /**
-   * Creates a depth-first search (DFS) visitor with a post-order traversal strategy
-   * using a queue as the journal to record visited elements.
-   *
-   * @param fulfill a function that takes an element of type `K` and returns an appropriate value of type `Option[V]`
-   * @param children a function that takes an element of type `K` and returns its sequence of child elements to traverse
-   * @return a new `DfsVisitorMapped[K, V]` instance configured for post-order traversal using a queue
-   */
-  def createPostMapJournal[K, V](fulfill: K => Option[V], children: K => Seq[K]): DfsVisitorMapped[K, V] =
-    create(Post, MapJournal.empty, fulfill, children)
-
-  /**
-   * Creates a new `DfsVisitorMapped` instance configured with a pre-order stack-based traversal strategy.
-   *
-   * This method initializes the depth-first search traversal with a pre-order traversal approach,
-   * where the stack is used as the underlying data structure to manage visited elements.
-   * The provided functions define how each element is mapped to an associated value and the children
-   * of each element to traverse recursively.
-   *
-   * @param fulfill a function that takes an element of type `K` and returns an appropriate value of type `Option[V]`
-   * @param children a function that takes an element of type `K` and returns its sequence of child elements to traverse
-   * @return a new `DfsVisitorMapped[K, V]` instance configured for pre-order stack-based depth-first traversal
-   */
-  def createPreFunctionMapJournal[K, V](fulfill: K => Option[V], children: K => Seq[K]): DfsVisitorMapped[K, V] =
-    create(Pre, FunctionMapJournal.empty(fulfill), fulfill, children)
-
-  /**
-   * Creates a `DfsVisitorMapped` instance configured for depth-first traversal using a post-visit stack-based strategy.
-   *
-   * This method initializes a `DfsVisitorMapped` with a `Post` phase, an empty journal, and the specified
-   * transformation and child extraction functions. It is tailored for managing depth-first traversal
-   * where the post-visit operation is applied after visiting all the child elements of a node.
-   *
-   * @param fulfill a function that takes an element of type `K` and returns an appropriate value of type `Option[V]`
-   * @param children a function that takes an element of type `K` and returns its sequence of child elements to traverse
-   * @return a new instance of `DfsVisitorMapped[K, V]` configured with the post-visit traversal strategy
-   */
-  def createPostFunctionMapJournal[K, V](fulfill: K => Option[V], children: K => Seq[K]): DfsVisitorMapped[K, V] =
-    create(Post, FunctionMapJournal.empty(fulfill), fulfill, children)
-
-  /**
-   * Performs recursive operations on a key-value pair with a depth-first search (DFS) visitor,
-   * applying in-visit logic if the provided sequence of children contains at most two elements.
-   * Depending on the contents of the sequence `xs`, it updates the visitor state accordingly.
-   *
-   * @param visitor the initial DFS visitor instance to be updated during the recursive process
-   * @param k       the current key to be processed
-   * @param v       the current associated value of the key
-   * @param xs      a sequence of child keys to be processed; must contain at most two elements
-   */
-  private def recurseWithInVisit[K, V](visitor: DfsVisitorMapped[K, V], k: K, v: Option[V], xs: Seq[K]) = {
-    require(xs.length <= 2, "xs must contain at most two elements")
-    val visitor1: DfsVisitorMapped[K, V] = xs.headOption.fold(visitor)(visitor.dfs)
-    val visitor2: DfsVisitorMapped[K, V] = visitor1.visit(In)(k -> v)
-    xs.lastOption.fold(visitor2)(visitor2.dfs)
-  }
-}
-
-/**
- * A specialized visitor class that implements depth-first search (DFS) traversal while maintaining
- * the state of the traversal using a map of messages to corresponding appendable data structures.
- *
- * This class extends the behavior of `AbstractVisitorMappedWithChildren` and `Dfs`, enabling traversal
- * of hierarchical or graph-based data structures. It processes elements and their associated dependencies
- * recursively, applying pre-visit, in-visit, self-visit, and post-visit phases during traversal.
- *
- * The traversal process uses key-value pairs where keys represent the primary elements and optional values
- * represent associated data to be processed. This class supports updates to the visitor state and provides
- * flexible configuration of child element lookups and dependency resolution.
- *
- * @tparam K the type representing the primary elements to be traversed
- * @tparam V the type representing the associated data for elements in the traversal
- * @constructor Creates a new instance of `DfsOriginVisitor` with the specified map, children function,
- *              and unmap function.
- * @param map      a mapping from `Message` to `Appendable[(K, Option[V])]`, representing the current
- *                 state of the visitor and its associated data
- * @param children a function that takes an element of type `K` and returns its sequence of child elements of type `V`
- * @param unmap    a function that "unmaps" a value of type `V` to an appropriate element of type `K`
- */
-case class DfsOriginVisitor[K, V](map: Map[Message, Appendable[(K, Option[V])]], children: K => Seq[V], unmap: V => K) extends
-  AbstractVisitorMappedWithChildren[K, V, V](map, vo => _ => vo, children) with Dfs[K, DfsOriginVisitor[K, V]] {
-
-  /**
-   * Processes an element of type `K` with an optional value of type `V` using a depth-first traversal strategy.
-   * This method applies pre-visit, in-visit (conditional on specific criteria), self-visit, and post-visit stages
-   * during the traversal and updates the state of the `DfsOriginVisitor`.
-   *
-   * @param k  the element of type `K` to be processed during the traversal.
-   * @param vo an optional value of type `V` associated with the element `h`.
-   * @return a new instance of `DfsOriginVisitor[K, V]` containing the updated state after processing
-   *         the provided element and its dependencies.
-   */
-  def dfsOrigin(k: K, vo: Option[V]): DfsOriginVisitor[K, V] = {
-    val kv: (K, Option[V]) = k -> vo
-
-    def performRecursion(xs: Seq[(K, Option[V])], visitor: DfsOriginVisitor[K, V]) =
-      if (xs.length <= 2 && map.contains(In))
-        DfsOriginVisitor.recurseWithInVisit(visitor, kv, xs)
-      else
-        xs.foldLeft[DfsOriginVisitor[K, V]](visitor)((accum, z) => accum.dfsOrigin.tupled(z))
-
-    // First do a pre-visit, then a "SelfVisit", next perform the recursion, finally do a post-visit
-    performRecursion(children(k) map (z => (unmap(z), Some(z))), visit(Pre)(kv).visit(SelfVisit)(kv)).visit(Post)(kv)
-  }
-
-  /**
-   * Recursively processes an element of type `K` using a depth-first traversal strategy.
-   * This method updates the internal state of the `DfsVisitor` at each step of the recursion,
-   * applying pre-, in-, and post-visit modifications as determined by the `Pre`-, `In`- and `Post`-messages.
-   *
-   * @param x the element of type `K` to be processed and traversed recursively.
-   * @return a new instance of `DfsVisitor` with the updated state after processing
-   *         and visiting the provided element and its sub-elements.
-   */
-  def dfs(x: K): DfsOriginVisitor[K, V] =
-    dfsOrigin(x, None)
-
-  /**
-   * Make a visit, with the given message and `X` value, on this `Visitor` and return a new `Visitor`.
-   *
-   * This method defines the behavior for handling a `Message` in the context
-   * of the Visitor pattern. The implementation of this method should use the provided
-   * message and state to determine the next state and return the appropriate `Visitor`.
-   *
-   * @param msg the message to be processed by the visitor
-   * @param kv  the current state or context associated with the visitor
-   * @return a new `Visitor[X]` instance that represents the updated state after processing the message
-   */
-  override def visit(msg: Message)(kv: (K, Option[V])): DfsOriginVisitor[K, V] =
-    super.visit(msg)(kv).asInstanceOf[DfsOriginVisitor[K, V]]
-
-  /**
-   * Creates a new `Visitor` instance with the provided updated mapAppendables.
-   *
-   * This method is used to update the internal state of the Visitor by creating
-   * a new instance with the modified mappings from `Message` to `Appendable`.
-   *
-   * @param map a map containing updated associations of `Message` to `Appendable[K]`
-   * @return a new `Visitor[K]` instance that reflects the updated mapAppendables
-   */
-  def unit(map: Map[Message, Appendable[(K, Option[V])]]): DfsOriginVisitor[K, V] =
-    copy(map = map)
-}
-
-/**
- * Companion object for the `DfsOriginVisitor` class, providing factory methods
- * for creating instances of `DfsOriginVisitor` with different configurations.
- */
-object DfsOriginVisitor {
-
-  /**
-   * Creates a new instance of `DfsVisitorMapped` to manage depth-first traversal using a visitor pattern.
-   *
-   * @param message  the `Message` instance representing the context or phase of the traversal
-   * @param journal  the `Appendable` structure for recording visited elements as pairs, each having type `(K, V)`
-   * @param children a function that takes an element of type `K` and returns its sequence of child elements to traverse
-   * @return a new instance of `DfsVisitorMapped[K, V]` configured with the specified message, appendable structure, and functions
-   */
-  def create[K, V](message: Message, journal: AbstractMapJournal[K, Option[V]], children: K => Seq[V], h: V => K): DfsOriginVisitor[K, V] =
-    new DfsOriginVisitor[K, V](Map(message -> journal), children, h)
-
-  /**
-   * Creates a `DfsVisitorMapped` instance configured for depth-first traversal using a pre-visit phase
-   * and a queue-based mechanism for recording visited elements.
-   *
-   * This method sets up the traversal with a pre-order strategy, utilizing a queue as the journal
-   * to manage visited elements during the traversal process. The functions provided define the mapping
-   * of an element to a value and how the children of each element are determined for recursive traversal.
-   *
-   * @param children a function that takes an element of type `K` and returns its sequence of child elements to traverse
-   * @return a new `DfsVisitorMapped[K, V]` instance configured for pre-order traversal using a queue
-   */
-  def createPreMapJournal[K, V](children: K => Seq[V], h: V => K): DfsOriginVisitor[K, V] =
-    create(Pre, MapJournal.empty, children, h)
-
-  /**
-   * Creates a depth-first search (DFS) visitor with a post-order traversal strategy
-   * using a queue as the journal to record visited elements.
-   *
-   * @param children a function that takes an element of type `K` and returns its sequence of child elements to traverse
-   * @return a new `DfsVisitorMapped[K, V]` instance configured for post-order traversal using a queue
-   */
-  def createPostMapJournal[K, V](children: K => Seq[V], h: V => K): DfsOriginVisitor[K, V] =
-    create(Post, MapJournal.empty, children, h)
-
-  /**
-   * Performs a recursive traversal with specific handling for "In" messages.
-   * This method processes the provided element `kv` within a recursive structure
-   * and updates the visitor state accordingly, ensuring it supports the case
-   * of sequences with at most two elements.
-   *
-   * CONSIDER should we pass in the dfs function as a parameter?
-   *
-   * @param visitor the current `DfsVisitor` instance that maintains
-   *                the state of the recursive traversal.
-   * @param x       the element of type `X` that is being processed and traversed.
-   * @param xs      a sequence of type `X` representing the sub-elements related to `kv`.
-   *                This sequence must contain at most two elements.
-   */
-  private def recurseWithInVisit[X, Y](visitor: DfsOriginVisitor[X, Y], x: (X, Option[Y]), xs: Seq[(X, Option[Y])]) = {
-    require(xs.length <= 2, "xs must contain at most two elements")
-    val visitor1 = xs.headOption.fold(visitor)(visitor.dfsOrigin)
-    val visitor2 = visitor1.visit(In)(x)
-    xs.lastOption.fold(visitor2)(visitor2.dfsOrigin)
-  }
-}
