+++ src/main/scala/com/phasmidsoftware/visitor/misc/NonAppendable.scala
@@ -1,4 +1,4 @@
-package com.phasmidsoftware.visitor
+package com.phasmidsoftware.visitor.misc
 
 /**
  * A case class implementation of the `Appendable` trait which executes a provided function
@@ -12,7 +12,7 @@ package com.phasmidsoftware.visitor
  * @param f a function to be executed with each appended element
  * @tparam X the type of elements that can be appended to this `Appendable`
  */
-case class NonAppendable[X](f: X => Unit) extends Appendable[X] {
+case class NonAppendable[X](f: X => Unit) extends com.phasmidsoftware.visitor.core.Appendable[X] {
 
   /**
    * "Appends" the specified element to this `Appendable` object, returning a new instance
@@ -23,7 +23,7 @@ case class NonAppendable[X](f: X => Unit) extends Appendable[X] {
    * @param x the element to be "appended"
    * @return a new `Appendable[X]` instance containing the existing elements and the newly appended element
    */
-  def append(x: X): Appendable[X] = {
+  def append(x: X): com.phasmidsoftware.visitor.core.Appendable[X] = {
     f(x)
     this
   }
