+++ src/test/scala/com/phasmidsoftware/visitor/core/TranslatedSpec.scala
@@ -1,6 +1,6 @@
 package com.phasmidsoftware.visitor.core
 
-import com.phasmidsoftware.visitor.{FunctionMapJournal, MapJournal}
+import com.phasmidsoftware.visitor.misc.{FunctionMapJournal, MapJournal}
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
